fix: Assume localhost server is always available
The navigator.onLine tells nothing about if the server can be reached but tells if a network connection is available. It really should not be used at all for this state indicator but seems to work reasonably well in all other cases except when the server is running on localhost and you do not have a network connection.

Fixes https://github.com/vaadin/flow/issues/13616